requirements.txt
Flask
gunicorn
pymongo>=4.7
python-dotenv
flask-jwt-extended
flask-bcrypt
pytest
requests
celery
redis
minio
PyMuPDF